requirements.txt
# Core
Flask>=2.3.0
requests>=2.31.0

# Security
bcrypt>=4.0.0

# Payments
stripe>=7.0.0

# Database (PostgreSQL support for production scale)
psycopg2-binary>=2.9.0

# Caching (reduce API calls)
cachetools>=5.3.0

# Rate limiting
flask-limiter>=3.5.0

# Kelion Super AI (Claude Integration)
anthropic>=0.18.0
python-dotenv>=1.0.0
